+++ python/hokku/screens/seeedstudio_e1004/__init__.py
@@ -34,6 +34,9 @@ SPEC = Esp32Spec(
     nvs_size=constants.NVS_SIZE,
     app_offset=constants.APP_OFFSET,
     bootloader_offset=constants.BOOTLOADER_OFFSET,
+    ota1_offset=constants.OTA1_OFFSET,
+    otadata_offset=constants.OTADATA_OFFSET,
+    otadata_size=constants.OTADATA_SIZE,
     vid=constants.ESP32S3_VID,
     pid=constants.ESP32S3_PID,
     baud=constants.ESPTOOL_BAUD,
@@ -64,9 +67,13 @@ list_serial_ports = partial(_device.list_serial_ports, SPEC)
 read_device_flash = partial(_device.read_device_flash, SPEC)
 parse_device_state = partial(_device.parse_device_state, SPEC)
 scan_devices = partial(_device.scan_devices, SPEC)
+active_ota_slot = _device.active_ota_slot  # pure otadata parse, takes no spec
+boot_app = partial(_device.boot_app, SPEC)
 
 # Flashing (spec-bound).
 flash_firmware = partial(_flasher.flash_firmware, SPEC)
+describe_flash_parts = partial(_flasher.describe_flash_parts, SPEC)
+describe_config_part = partial(_flasher.describe_config_part, SPEC)
 write_config = partial(_flasher.write_config, SPEC)
 flash_device = partial(_flasher.flash_device, SPEC)
 
@@ -80,10 +87,14 @@ __all__ = [
     "EsptoolError",
     "NvsToolUnavailable",
     "SeeedstudioE1004Display",
+    "active_ota_slot",
     "app_image_from_file",
+    "boot_app",
     "build_nvs_binary",
     "bundled_firmware_version",
     "constants",
+    "describe_config_part",
+    "describe_flash_parts",
     "flash_device",
     "flash_firmware",
     "list_firmware_files",